+++ CHANGELOG.md
@@ -3,7 +3,11 @@
     - The errors will now appear as tmux messages
 - `Change` Perform selection from the first page instead of the last when text longer than the screen is piped into mless
 - `Feature` It is now possible to define multiple modes in the config file and switch between them by pressing `<Space>` followed by the appropriate hotkey
-- `Feature` Colors of hints during selection can now be customized in the config file
+- `Feature` Colors of hints displayed during selection can now be customized in the config file
+- `Feature` It is now possible to customize the following aspects of the mode switching dialog:
+    - Divider color
+    - Mode hotkeys' color
+    - Mode names' color
 
 # v0.1.0
 First release, no changelog was kept